requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
drf-spectacular>=0.27,<0.28
django-cors-headers>=4.4,<4.5
psycopg2-binary>=2.9,<3.0
PyJWT>=2.9,<3.0
